refactor(dogstatsd): migrate server component to v2 architecture
Migrates comp/dogstatsd/server from V1 to V2 component architecture:
- component.go + params.go moved to def/
- All implementation files (server.go, batch.go, enrich.go, parse*.go, etc.) moved to impl/
- server_mock.go moved to mock/mock.go
- New fx/fx.go with Module(params Params) using fxutil.ProvideComponentConstructor
- fx.In/Lifecycle/Hook replaced with compdef.In/Lifecycle/Hook in impl/server.go
- Provides struct with compdef.Out instead of fx.Out
- newServer renamed to NewComponent (exported) with Provides return type
- All callers updated to import from server/def instead of server
- Test utility updated to use serverdef.Component and fxutil.ProvideComponentConstructor
