+++ comp/dogstatsd/bundle.go
@@ -8,7 +8,8 @@ package dogstatsd //nolint:revive // TODO(AML) Fix revive linter
 import (
 	pidmapfx "github.com/DataDog/datadog-agent/comp/dogstatsd/pidmap/fx"
 	replayfx "github.com/DataDog/datadog-agent/comp/dogstatsd/replay/fx"
-	"github.com/DataDog/datadog-agent/comp/dogstatsd/server"
+	server "github.com/DataDog/datadog-agent/comp/dogstatsd/server/def"
+	serverfx "github.com/DataDog/datadog-agent/comp/dogstatsd/server/fx"
 	serverdebugfx "github.com/DataDog/datadog-agent/comp/dogstatsd/serverDebug/fx"
 	statusfx "github.com/DataDog/datadog-agent/comp/dogstatsd/status/fx"
 	offlinereporterfx "github.com/DataDog/datadog-agent/comp/offlinereporter/fx"
@@ -24,6 +25,6 @@ func Bundle(params server.Params) fxutil.BundleOptions {
 		serverdebugfx.Module(),
 		replayfx.Module(),
 		pidmapfx.Module(),
-		server.Module(params),
+		serverfx.Module(params),
 		statusfx.Module())
 }
