+++ comp/dogstatsd/bundle_mock.go
@@ -10,7 +10,7 @@ package dogstatsd
 import (
 	pidmapfx "github.com/DataDog/datadog-agent/comp/dogstatsd/pidmap/fx"
 	replayfx "github.com/DataDog/datadog-agent/comp/dogstatsd/replay/fx"
-	"github.com/DataDog/datadog-agent/comp/dogstatsd/server"
+	servermock "github.com/DataDog/datadog-agent/comp/dogstatsd/server/mock"
 	serverdebugmock "github.com/DataDog/datadog-agent/comp/dogstatsd/serverDebug/mock"
 	statsdimpl "github.com/DataDog/datadog-agent/comp/dogstatsd/statsd/impl"
 	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
@@ -20,7 +20,7 @@ import (
 func MockBundle() fxutil.BundleOptions {
 	return fxutil.Bundle(
 		serverdebugmock.MockModule(),
-		server.MockModule(),
+		servermock.MockModule(),
 		replayfx.Module(),
 		pidmapfx.Module())
 }
