+++ comp/dogstatsd/status/impl/status_test.go
@@ -16,7 +16,7 @@ import (
 	"github.com/stretchr/testify/assert"
 
 	// We need to include this to make sure the Dogstatsd expvars are initialized
-	_ "github.com/DataDog/datadog-agent/comp/dogstatsd/server"
+	_ "github.com/DataDog/datadog-agent/comp/dogstatsd/server/impl"
 )
 
 func TestStatusDisabledWhenADPEnabled(t *testing.T) {
