+++ pkg/jmxfetch/runner.go
@@ -13,7 +13,7 @@ import (
 	"github.com/DataDog/datadog-agent/comp/agent/jmxlogger"
 	"github.com/DataDog/datadog-agent/comp/core/autodiscovery/integration"
 	ipc "github.com/DataDog/datadog-agent/comp/core/ipc/def"
-	dogstatsdServer "github.com/DataDog/datadog-agent/comp/dogstatsd/server"
+	dogstatsdServer "github.com/DataDog/datadog-agent/comp/dogstatsd/server/def"
 	pkgconfigsetup "github.com/DataDog/datadog-agent/pkg/config/setup"
 	jmxStatus "github.com/DataDog/datadog-agent/pkg/status/jmx"
 )
